adding more tests for hplsql parser
